+++ Assets/Scripts/GameStart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 38caa907a0b13dd49b881ff2a47adeee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
